Tcp server and upate run configurations

+++ SentToDo.Deskband/InstallDeskband.ps1
@@ -1,9 +1,13 @@
 ﻿#Requires -RunAsAdministrator
 
+param (
+    [switch]$uninstall = $uninstall
+)
+
 $Regasm = $Env:SystemRoot + "\Microsoft.NET\Framework64\v4.0.30319\regasm.exe"
 $NewFile = "SentToDo.Deskband\bin\Debug\SentToDo.Deskband.dll"
 $BuildPath = Split-Path -Path $NewFile
-$ActivatedFile = $BuildPath + "SentToDoActivated.dll"
+$ActivatedFile = $BuildPath + "\SentToDoActivated.dll"
 
 if (Test-Path $ActivatedFile)
 {
@@ -12,7 +16,10 @@ if (Test-Path $ActivatedFile)
     Remove-Item $ActivatedFile
 }
 
-Copy-Item -Path $NewFile -Destination $ActivatedFile -force
-.$Regasm /nologo /codebase $ActivatedFile
-Stop-Process -ProcessName explorer
-Start-Process explorer
+if (-Not $uninstall)
+{
+    Copy-Item -Path $NewFile -Destination $ActivatedFile -force
+    .$Regasm /nologo /codebase $ActivatedFile
+    Stop-Process -ProcessName explorer
+    Start-Process explorer
+}
